change version slector position
