+++ site/theme/static/header.less
@@ -62,10 +62,15 @@
   box-shadow: none;
 }
 
-.lang {
+.lang,
+.version {
   float: right;
   margin-top: 29px;
   margin-left: 12px;
+}
+
+.version {
+  width: 68px;
 }
 
 #nav {
